fix: introduce modern typings for TS >=4.1
