+++ change/@griffel-style-types-88cd8fa1-371b-4c2e-bb29-0d88568c620d.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "fix: resolve recursive typings issue",
+  "packageName": "@griffel/style-types",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
