+++ e2e/typescript/src/assets/fixture.ts
@@ -62,6 +62,13 @@ assertType({
   ':hover': { '--color': 'red' },
 });
 
+assertType({
+  ':hover': { content: "'foo'" },
+});
+assertType({
+  ':hover': { animationName: { from: {}, to: {} } },
+});
+
 // Custom selectors
 //
 
@@ -86,6 +93,9 @@ assertType({
 
 assertType({
   '.bar': { '--color': 'red' },
+});
+assertType({
+  '.bar': { animationName: { from: {}, to: {} } },
 });
 
 // Nested custom selectors
@@ -192,7 +202,6 @@ assertType({
   '@media screen and (max-width: 992px)': {
     ...typedMixin,
   },
-  // @ts-expect-error FIXME - this is supposed to pass!
   '@media(forced-colors: active)': {
     ...typedMixin,
     color: 'var(--customColor)',
@@ -201,6 +210,13 @@ assertType({
     ...typedMixin,
   },
   ':before': {
+    ...typedMixin,
+    color: 'var(--customColor)',
+  },
+  '& .foo': {
+    ...typedMixin,
+  },
+  '& .bar': {
     ...typedMixin,
     color: 'var(--customColor)',
   },
@@ -231,31 +247,32 @@ assertType({
 //
 
 assertType({
-  // @ts-expect-error "1" is invalid value for "overflowX"
   ':hover:focus': {
+    // @ts-expect-error "1" is invalid value for "overflowX"
     overflowX: '1',
   },
 });
 
 assertType({
-  // @ts-expect-error "padding" is banned
   ':hover:focus': {
+    // @ts-expect-error "padding" is banned
     padding: 0,
   },
 });
 
 assertType({
-  // @ts-expect-error "paddingLeft" cannot be numeric value
   ':hover:focus': {
+    // @ts-expect-error "5" is an invalid value for paddingLeft
     paddingLeft: 5,
   },
 });
 
 assertType({
-  // @ts-expect-error "1" is invalid value for "overflowX", padding is banned, paddingLeft cannot be a numeric value
   ':hover:focus': {
     overflowX: 'scroll',
+    // @ts-expect-error "padding" is banned
     padding: 0,
+    // @ts-expect-error "5" is an invalid value for paddingLeft
     paddingLeft: 5,
   },
 });
@@ -264,12 +281,14 @@ assertType({
   ':hover:focus': {
     // @ts-expect-error Object is not assignable to CSS property
     zIndex: { color: 'red' },
-    opacity: { color: 'red' }, // < no error here, TS only reports the first error in this case
+    // @ts-expect-error Object is not assignable to CSS property
+    opacity: { color: 'red' },
   },
   ':hover:active': {
     // @ts-expect-error Object is not assignable to CSS property
     opacity: { color: 'red' },
-    zIndex: { color: 'red' }, // < no error here, TS only reports the first error in this case
+    // @ts-expect-error Object is not assignable to CSS property
+    zIndex: { color: 'red' },
   },
 });
 
@@ -277,33 +296,37 @@ assertType({
 //
 
 assertType({
-  // @ts-expect-error "1" is invalid value for "overflowX", padding is banned, paddingLeft cannot be a numeric value
   '.foo': {
     '.baz': {
+      // @ts-expect-error outline-box is an invalid value for box-sizing
       overflowX: '1',
+      // @ts-expect-error "padding" is banned
       padding: 0,
+      // @ts-expect-error "5" is an invalid value for paddingLeft
       paddingLeft: 5,
     },
   },
 });
 assertType({
-  // @ts-expect-error outline-box is an invalid value for box-sizing
   '.foo': {
+    // @ts-expect-error outline-box is an invalid value for box-sizing
     boxSizing: 'outline-box',
 
     '.bar': {
-      boxSizing: 'outline-box', // < no error here, TS only reports the error for the whole object
+      // @ts-expect-error outline-box is an invalid value for box-sizing
+      boxSizing: 'outline-box',
     },
   },
 });
 assertType({
-  // @ts-expect-error outline-box is an invalid value for box-sizing
   '.foo': {
+    // @ts-expect-error outline-box is an invalid value for box-sizing
     boxSizing: 'outline-box',
     zIndex: 1,
 
     '.bar': {
-      boxSizing: 'outline-box', // < no error here, TS only reports the error for the whole object
+      // @ts-expect-error outline-box is an invalid value for box-sizing
+      boxSizing: 'outline-box',
       zIndex: 1,
     },
   },
@@ -312,11 +335,14 @@ assertType({
   '.foo': {
     // @ts-expect-error Object is not assignable to CSS property
     zIndex: { color: 'red' },
-    opacity: { color: 'red' }, // < no error here, TS only reports the first error in this case
+    // @ts-expect-error Object is not assignable to CSS property
+    opacity: { color: 'red' },
 
     '.bar': {
-      zIndex: { color: 'red' }, // < no error here, TS only reports the first error in this case
-      opacity: { color: 'red' }, // < no error here, TS only reports the first error in this case
+      // @ts-expect-error Object is not assignable to CSS property
+      zIndex: { color: 'red' },
+      // @ts-expect-error Object is not assignable to CSS property
+      opacity: { color: 'red' },
     },
   },
 });
@@ -342,8 +368,11 @@ assertType({
     2,
     '2px',
   ],
-  // @ts-expect-error "paddingLeft" cannot be numeric value
   ':hover:active': {
-    paddingLeft: ['2px', 2],
+    paddingLeft: [
+      '2px',
+      // @ts-expect-error "paddingLeft" cannot be numeric value
+      2,
+    ],
   },
 });
