+++ e2e/typescript/src/assets/legacy/tsconfig.fixture.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "lib": ["ES2019", "dom"],
+    "isolatedModules": true,
+    "moduleResolution": "node",
+    "types": []
+  },
+  "include": ["*.ts"]
+}
