+++ e2e/typescript/src/test.ts
@@ -2,17 +2,24 @@ import { configureYarn, copyAssets, createTempDir, installPackages, packLocalPac
 import * as logSymbols from 'log-symbols';
 import * as path from 'path';
 
-async function performTest(tsVersion: string) {
+async function performTest(tsVersion: string, options: { mode?: 'legacy' | 'modern' } = {}) {
   let tempDir: string;
   let tscBin: string;
 
+  const { mode = 'modern' } = options;
+
   try {
     const rootDir = path.resolve(__dirname, '..', '..', '..');
+    const assetsPath = path.resolve(__dirname, 'assets', mode === 'legacy' ? 'legacy' : '');
 
     tempDir = createTempDir('typescript');
 
+    if (mode === 'legacy') {
+      console.log(logSymbols.info, 'Using legacy mode...');
+    }
+
     await copyAssets({
-      assetsPath: path.resolve(__dirname, 'assets'),
+      assetsPath: assetsPath,
       tempDir,
       renames: { 'tsconfig.fixture.json': 'tsconfig.json' },
     });
@@ -43,7 +50,10 @@ async function performTest(tsVersion: string) {
   }
 
   try {
-    await sh(`node ${tscBin} --noEmit --pretty`, tempDir);
+    const command = `node ${tscBin} --noEmit --pretty`;
+
+    console.log(logSymbols.info, 'Running', command);
+    await sh(command, tempDir);
 
     console.log(logSymbols.success, `Example project was successfully built with typescript@${tsVersion}`);
     console.log('');
@@ -65,7 +75,7 @@ async function performTest(tsVersion: string) {
 }
 
 (async () => {
-  await performTest('3.9');
+  await performTest('3.9', { mode: 'legacy' });
   await performTest('4.1');
   await performTest('4.4');
   await performTest('4.9');
