+++ packages/style-types/src/legacy/README.md
@@ -1,0 +1,1 @@
+This folder contains types for legacy version of TypeScript (TS <4).
