+++ packages/style-types/src/legacy/index.ts
@@ -1,0 +1,6 @@
+export type { GriffelAnimation, GriffelStyle } from './makeStyles';
+export type { GriffelResetStyle } from './makeResetStyles';
+export type { GriffelStaticStyle, GriffelStaticStyles } from './makeStaticStyles';
+
+export type { GriffelStylesCSSValue, ValueOrArray } from '../shared';
+export type { GriffelStylesUnsupportedCSSProperties } from '../unsupported-properties';
