+++ packages/style-types/src/legacy/makeStaticStyles.ts
@@ -1,0 +1,27 @@
+import * as CSS from 'csstype';
+
+//
+// Types for makeStaticStyles()
+// ---
+//
+
+export type GriffelStaticStyle = {
+  [key: string]: CSS.Properties &
+    // TODO Questionable: how else would users target their own children?
+    Record<string, any>; // eslint-disable-line @typescript-eslint/no-explicit-any
+} & {
+  '@font-face'?: {
+    fontFamily: string;
+    src: string;
+
+    fontFeatureSettings?: string;
+    fontStretch?: string;
+    fontStyle?: string;
+    fontVariant?: string;
+    fontVariationSettings?: string;
+    fontWeight?: number | string;
+
+    unicodeRange?: string;
+  };
+};
+export type GriffelStaticStyles = GriffelStaticStyle | string;
